requirements.txt
# Function dependencies, for example:
# package>=version
yfinance
pandas
google-cloud-bigquery
python-csv
requests
DateTime
regex
pyarrow
